onboarding trial/maverick-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.36.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-dotenv==1.0.1
pydantic==2.11.3
pydantic-settings==2.9.1
python-multipart==0.0.12
email-validator==2.2.0
